% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_phasic}
\alias{install_phasic}
\title{Install Python phasic backend}
\usage{
install_phasic(method = "auto", conda = "auto", pip = TRUE, ...)
}
\arguments{
\item{method}{Installation method to use. Default is "auto" which automatically
finds a suitable Python installation. Other options include "virtualenv" and "conda".}

\item{conda}{Path to conda executable. Only used when method = "conda".}

\item{pip}{Logical, use pip for installation (default TRUE).}

\item{...}{Additional arguments passed to \code{\link[reticulate]{py_install}}.}
}
\value{
NULL (invisibly). Called for side effect of installing Python package.
}
\description{
Helper function to install the Python phasic package required by this R package.
}
\examples{
\dontrun{
# Install phasic Python package
install_phasic()

# Install in a specific conda environment
install_phasic(method = "conda", conda = "/path/to/conda")
}

}
